+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneDiskEncryptionWindows10.Tests.ps1
@@ -1,0 +1,541 @@
+[CmdletBinding()]
+param(
+)
+$M365DSCTestFolder = Join-Path -Path $PSScriptRoot `
+                        -ChildPath '..\..\Unit' `
+                        -Resolve
+$CmdletModule = (Join-Path -Path $M365DSCTestFolder `
+            -ChildPath '\Stubs\Microsoft365.psm1' `
+            -Resolve)
+$GenericStubPath = (Join-Path -Path $M365DSCTestFolder `
+    -ChildPath '\Stubs\Generic.psm1' `
+    -Resolve)
+Import-Module -Name (Join-Path -Path $M365DSCTestFolder `
+        -ChildPath '\UnitTestHelper.psm1' `
+        -Resolve)
+
+$Global:DscHelper = New-M365DscUnitTestHelper -StubModule $CmdletModule `
+    -DscResource "IntuneDiskEncryptionWindows10" -GenericStubModule $GenericStubPath
+Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
+    InModuleScope -ModuleName $Global:DscHelper.ModuleName -ScriptBlock {
+        Invoke-Command -ScriptBlock $Global:DscHelper.InitializeScript -NoNewScope
+        BeforeAll {
+
+            $secpasswd = ConvertTo-SecureString (New-Guid | Out-String) -AsPlainText -Force
+            $Credential = New-Object System.Management.Automation.PSCredential ('tenantadmin@mydomain.com', $secpasswd)
+
+            Mock -CommandName Confirm-M365DSCDependencies -MockWith {
+            }
+
+            Mock -CommandName Get-PSSession -MockWith {
+            }
+
+            Mock -CommandName Remove-PSSession -MockWith {
+            }
+
+            Mock -CommandName New-MgBetaDeviceManagementConfigurationPolicy -MockWith {
+                return @{
+                    Id = '619bd4a4-3b3b-4441-bd6f-3f4c0c444870'
+                }
+            }
+
+            Mock -CommandName Remove-MgBetaDeviceManagementConfigurationPolicy -MockWith {
+            }
+
+            Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicy -MockWith {
+            }
+
+            Mock -CommandName Update-IntuneDeviceConfigurationPolicy -MockWith {
+            }
+
+            Mock -CommandName Get-IntuneSettingCatalogPolicySetting -MockWith {
+            }
+
+            Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicySetting -MockWith {
+                return ,@()
+            }
+
+            Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicyAssignment -MockWith {
+                return @(@{
+                    Id       = '12345-12345-12345-12345-12345'
+                    Source   = 'direct'
+                    SourceId = '12345-12345-12345-12345-12345'
+                    Target   = @{
+                        DeviceAndAppManagementAssignmentFilterId   = '12345-12345-12345-12345-12345'
+                        DeviceAndAppManagementAssignmentFilterType = 'none'
+                        AdditionalProperties                       = @(
+                            @{
+                                '@odata.type' = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                                groupId       = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                            }
+                        )
+                    }
+                })
+            }
+
+            Mock -CommandName Update-DeviceConfigurationPolicyAssignment -MockWith {
+            }
+
+            Mock -CommandName New-M365DSCConnection -MockWith {
+                return "Credentials"
+            }
+
+            # Mock Write-Host to hide output during the tests
+            Mock -CommandName Write-Host -MockWith {
+            }
+            $Script:exportedInstances =$null
+            $Script:ExportMode = $false
+        }
+        # Test contexts
+        Context -Name "The IntuneDiskEncryptionWindows10 should exist but it DOES NOT" -Fixture {
+            BeforeAll {
+                $testParams = @{
+                    Assignments = [CimInstance[]]@(
+                        (New-CimInstance -ClassName MSFT_DeviceManagementConfigurationPolicyAssignments -Property @{
+                            DataType     = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                            groupId = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                        } -ClientOnly)
+                    )
+                    Description = "FakeStringValue"
+                    DisplayName = "FakeStringValue"
+                    Id = "FakeStringValue"
+                    IdentificationField_Name = "1"
+                    IdentificationField = "Field"
+                    SecIdentificationField = "SecField"
+                    RoleScopeTagIds = @("FakeStringValue")
+                    Ensure = "Present"
+                    Credential = $Credential;
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicy -MockWith {
+                    return $null
+                }
+            }
+            It 'Should return Values from the Get method' {
+                (Get-TargetResource @testParams).Ensure | Should -Be 'Absent'
+            }
+            It 'Should return false from the Test method' {
+                Test-TargetResource @testParams | Should -Be $false
+            }
+            It 'Should Create the group from the Set method' {
+                Set-TargetResource @testParams
+                Should -Invoke -CommandName New-MgBetaDeviceManagementConfigurationPolicy -Exactly 1
+            }
+        }
+
+        Context -Name "The IntuneDiskEncryptionWindows10 exists but it SHOULD NOT" -Fixture {
+            BeforeAll {
+                $testParams = @{
+                    Assignments = [CimInstance[]]@(
+                        (New-CimInstance -ClassName MSFT_DeviceManagementConfigurationPolicyAssignments -Property @{
+                            DataType     = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                            groupId = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                        } -ClientOnly)
+                    )
+                    Description = "FakeStringValue"
+                    DisplayName = "FakeStringValue"
+                    Id = "FakeStringValue"
+                    IdentificationField_Name = "1"
+                    IdentificationField = "Field"
+                    SecIdentificationField = "SecField"
+                    RoleScopeTagIds = @("FakeStringValue")
+                    Ensure = 'Absent'
+                    Credential = $Credential;
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicy -MockWith {
+                    return @{
+                        Description = "FakeStringValue"
+                        Id = "FakeStringValue"
+                        Name = "FakeStringValue"
+                        RoleScopeTagIds = @("FakeStringValue")
+                        TemplateReference = @{
+                            TemplateId = '46ddfc50-d10f-4867-b852-9434254b3bff_1'
+                        }
+                    }
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicySetting -MockWith {
+                    return @{
+                        Id                   = '0'
+                        SettingDefinitions   = @(
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield_identificationfield'
+                                Name = 'IdentificationField'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                }
+                            },
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield'
+                                Name = 'IdentificationField_Name'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                }
+                            },
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield_secidentificationfield'
+                                Name = 'SecIdentificationField'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                }
+                            }
+                        )
+                        SettingInstance      = @{
+                            SettingDefinitionId              = 'device_vendor_msft_bitlocker_identificationfield'
+                            SettingInstanceTemplateReference = @{
+                                SettingInstanceTemplateId = '3aeb9145-2c02-4086-8886-44dbe09c2f62'
+                            }
+                            AdditionalProperties             = @(
+                                @{
+                                    '@odata.type'               = '#microsoft.graph.deviceManagementConfigurationChoiceSettingInstance'
+                                    choiceSettingValue = @(
+                                        @{
+                                            children = @(
+                                                @{
+                                                    '@odata.type'       = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                    settingDefinitionId = 'device_vendor_msft_bitlocker_identificationfield_identificationfield'
+                                                    simpleSettingValue  = @{
+                                                        '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                        value = 'Field'
+                                                    }
+                                                },
+                                                @{
+                                                    '@odata.type'       = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                    settingDefinitionId = 'device_vendor_msft_bitlocker_identificationfield_secidentificationfield'
+                                                    simpleSettingValue  = @{
+                                                        '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                        value = 'SecField'
+                                                    }
+                                                }
+                                            )
+                                            value = 'device_vendor_msft_bitlocker_identificationfield_1'
+                                        }
+                                    )
+                                }
+                            )
+                        }
+                        AdditionalProperties = @{}
+                    }
+                }
+            }
+
+            It 'Should return Values from the Get method' {
+                (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
+            }
+
+            It 'Should return true from the Test method' {
+                Test-TargetResource @testParams | Should -Be $false
+            }
+
+            It 'Should Remove the group from the Set method' {
+                Set-TargetResource @testParams
+                Should -Invoke -CommandName Remove-MgBetaDeviceManagementConfigurationPolicy -Exactly 1
+            }
+        }
+        Context -Name "The IntuneDiskEncryptionWindows10 Exists and Values are already in the desired state" -Fixture {
+            BeforeAll {
+                $testParams = @{
+                    Assignments = [CimInstance[]]@(
+                        (New-CimInstance -ClassName MSFT_DeviceManagementConfigurationPolicyAssignments -Property @{
+                            DataType     = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                            groupId = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                        } -ClientOnly)
+                    )
+                    Description = "FakeStringValue"
+                    DisplayName = "FakeStringValue"
+                    Id = "FakeStringValue"
+                    IdentificationField_Name = "1"
+                    IdentificationField = "Field"
+                    SecIdentificationField = "SecField"
+                    RoleScopeTagIds = @("FakeStringValue")
+                    Ensure = 'Present'
+                    Credential = $Credential;
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicy -MockWith {
+                    return @{
+                        Description = "FakeStringValue"
+                        Id = "FakeStringValue"
+                        Name = "FakeStringValue"
+                        RoleScopeTagIds = @("FakeStringValue")
+                        TemplateReference = @{
+                            TemplateId = '46ddfc50-d10f-4867-b852-9434254b3bff_1'
+                        }
+                    }
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicySetting -MockWith {
+                    return @{
+                        Id                   = '0'
+                        SettingDefinitions   = @(
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield_identificationfield'
+                                Name = 'IdentificationField'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                }
+                            },
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield'
+                                Name = 'IdentificationField_Name'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                }
+                            },
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield_secidentificationfield'
+                                Name = 'SecIdentificationField'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                }
+                            }
+                        )
+                        SettingInstance      = @{
+                            SettingDefinitionId              = 'device_vendor_msft_bitlocker_identificationfield'
+                            SettingInstanceTemplateReference = @{
+                                SettingInstanceTemplateId = '3aeb9145-2c02-4086-8886-44dbe09c2f62'
+                            }
+                            AdditionalProperties             = @(
+                                @{
+                                    '@odata.type'               = '#microsoft.graph.deviceManagementConfigurationChoiceSettingInstance'
+                                    choiceSettingValue = @(
+                                        @{
+                                            children = @(
+                                                @{
+                                                    '@odata.type'       = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                    settingDefinitionId = 'device_vendor_msft_bitlocker_identificationfield_identificationfield'
+                                                    simpleSettingValue  = @{
+                                                        '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                        value = 'Field'
+                                                    }
+                                                },
+                                                @{
+                                                    '@odata.type'       = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                    settingDefinitionId = 'device_vendor_msft_bitlocker_identificationfield_secidentificationfield'
+                                                    simpleSettingValue  = @{
+                                                        '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                        value = 'SecField'
+                                                    }
+                                                }
+                                            )
+                                            value = 'device_vendor_msft_bitlocker_identificationfield_1'
+                                        }
+                                    )
+                                }
+                            )
+                        }
+                        AdditionalProperties = @{}
+                    }
+                }
+            }
+
+            It 'Should return true from the Test method' {
+                Test-TargetResource @testParams | Should -Be $true
+            }
+        }
+
+        Context -Name "The IntuneDiskEncryptionWindows10 exists and values are NOT in the desired state" -Fixture {
+            BeforeAll {
+                $testParams = @{
+                    Assignments = [CimInstance[]]@(
+                        (New-CimInstance -ClassName MSFT_DeviceManagementConfigurationPolicyAssignments -Property @{
+                            DataType     = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                            groupId = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                        } -ClientOnly)
+                    )
+                    Description = "FakeStringValue"
+                    Id = "FakeStringValue"
+                    IdentificationField_Name = "1"
+                    IdentificationField = "Field"
+                    SecIdentificationField = "SecField"
+                    DisplayName = "FakeStringValue"
+                    RoleScopeTagIds = @("FakeStringValue")
+                    Ensure = 'Present'
+                    Credential = $Credential;
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicy -MockWith {
+                    return @{
+                        Description = "FakeStringValue"
+                        Id = "FakeStringValue"
+                        Name = "FakeStringValue"
+                        RoleScopeTagIds = @("FakeStringValue")
+                        TemplateReference = @{
+                            TemplateId = '46ddfc50-d10f-4867-b852-9434254b3bff_1'
+                        }
+                    }
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicySetting -MockWith {
+                    return @{
+                        Id                   = '0'
+                        SettingDefinitions   = @(
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield_identificationfield'
+                                Name = 'IdentificationField'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                }
+                            },
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield'
+                                Name = 'IdentificationField_Name'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                }
+                            },
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield_secidentificationfield'
+                                Name = 'SecIdentificationField'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                }
+                            }
+                        )
+                        SettingInstance      = @{
+                            SettingDefinitionId              = 'device_vendor_msft_bitlocker_identificationfield'
+                            SettingInstanceTemplateReference = @{
+                                SettingInstanceTemplateId = '3aeb9145-2c02-4086-8886-44dbe09c2f62'
+                            }
+                            AdditionalProperties             = @(
+                                @{
+                                    '@odata.type'               = '#microsoft.graph.deviceManagementConfigurationChoiceSettingInstance'
+                                    choiceSettingValue = @(
+                                        @{
+                                            children = @(
+                                                @{
+                                                    '@odata.type'       = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                    settingDefinitionId = 'device_vendor_msft_bitlocker_identificationfield_identificationfield'
+                                                    simpleSettingValue  = @{
+                                                        '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                        value = 'ChangedValue' #drift
+                                                    }
+                                                },
+                                                @{
+                                                    '@odata.type'       = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                    settingDefinitionId = 'device_vendor_msft_bitlocker_identificationfield_secidentificationfield'
+                                                    simpleSettingValue  = @{
+                                                        '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                        value = 'SecField'
+                                                    }
+                                                }
+                                            )
+                                            value = 'device_vendor_msft_bitlocker_identificationfield_1'
+                                        }
+                                    )
+                                }
+                            )
+                        }
+                        AdditionalProperties = @{}
+                    }
+                }
+            }
+
+            It 'Should return Values from the Get method' {
+                (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
+            }
+
+            It 'Should return false from the Test method' {
+                Test-TargetResource @testParams | Should -Be $false
+            }
+
+            It 'Should call the Set method' {
+                Set-TargetResource @testParams
+                Should -Invoke -CommandName Update-IntuneDeviceConfigurationPolicy -Exactly 1
+            }
+        }
+
+        Context -Name 'ReverseDSC Tests' -Fixture {
+            BeforeAll {
+                $Global:CurrentModeIsExport = $true
+                $Global:PartialExportFileName = "$(New-Guid).partial.ps1"
+                $testParams = @{
+                    Credential = $Credential
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicy -MockWith {
+                    return @{
+                        Description = "FakeStringValue"
+                        Id = "FakeStringValue"
+                        Name = "FakeStringValue"
+                        RoleScopeTagIds = @("FakeStringValue")
+                        TemplateReference = @{
+                            TemplateId = '46ddfc50-d10f-4867-b852-9434254b3bff_1'
+                        }
+                    }
+                }
+
+                Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicySetting -MockWith {
+                    return @{
+                        Id                   = '0'
+                        SettingDefinitions   = @(
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield_identificationfield'
+                                Name = 'IdentificationField'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                }
+                            },
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield'
+                                Name = 'IdentificationField_Name'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                }
+                            },
+                            @{
+                                Id = 'device_vendor_msft_bitlocker_identificationfield_secidentificationfield'
+                                Name = 'SecIdentificationField'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                }
+                            }
+                        )
+                        SettingInstance      = @{
+                            SettingDefinitionId              = 'device_vendor_msft_bitlocker_identificationfield'
+                            SettingInstanceTemplateReference = @{
+                                SettingInstanceTemplateId = '3aeb9145-2c02-4086-8886-44dbe09c2f62'
+                            }
+                            AdditionalProperties             = @(
+                                @{
+                                    '@odata.type'               = '#microsoft.graph.deviceManagementConfigurationChoiceSettingInstance'
+                                    choiceSettingValue = @(
+                                        @{
+                                            children = @(
+                                                @{
+                                                    '@odata.type'       = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                    settingDefinitionId = 'device_vendor_msft_bitlocker_identificationfield_identificationfield'
+                                                    simpleSettingValue  = @{
+                                                        '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                        value = 'Field'
+                                                    }
+                                                },
+                                                @{
+                                                    '@odata.type'       = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                    settingDefinitionId = 'device_vendor_msft_bitlocker_identificationfield_secidentificationfield'
+                                                    simpleSettingValue  = @{
+                                                        '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                        value = 'SecField'
+                                                    }
+                                                }
+                                            )
+                                            value = 'device_vendor_msft_bitlocker_identificationfield_1'
+                                        }
+                                    )
+                                }
+                            )
+                        }
+                        AdditionalProperties = @{}
+                    }
+                }
+            }
+            It 'Should Reverse Engineer resource from the Export method' {
+                $result = Export-TargetResource @testParams
+                $result | Should -Not -BeNullOrEmpty
+            }
+        }
+    }
+}
+
+Invoke-Command -ScriptBlock $Global:DscHelper.CleanupScript -NoNewScope