+++ CHANGELOG.md
@@ -5,9 +5,15 @@
 * EXOHostedOutboundSpamFilterPolicy
   * Changed the RecipientLimitInternalPerHour, RecipientLimitPerDay, and
     RecipientLimitExternalPerHour parameters to UInt32.
+* IntuneDiskEncryptionWindows10
+  * Initial Release  
+    FIXES [#4050](https://github.com/microsoft/Microsoft365DSC/issues/4050)
 * M365DSCDRGUtil
+  * Fixes an issue with nested and duplicate settings in the settings catalog
   * Add support for converting Intune assignments directly from Graph.  
     FIXES [#4875](https://github.com/microsoft/Microsoft365DSC/issues/4875)
+* M365DSCResourceGenerator
+  * Update Intune resource generation.
 * M365DSCReport
   * Changes behaviour to not throw on empty configuration during report generation.  
     FIXES [#4559](https://github.com/microsoft/Microsoft365DSC/issues/4559)